--- a/net/icmp/Make.defs
+++ b/net/icmp/Make.defs
@@ -23,7 +23,7 @@ ifneq ($(CONFIG_NET_ICMP_NO_STACK),y)
 
 # ICMP source files
 
-NET_CSRCS += icmp_input.c
+NET_CSRCS += icmp_input.c icmp_reply.c
 
 ifeq ($(CONFIG_NET_ICMP_SOCKET),y)
 SOCK_CSRCS += icmp_sockif.c icmp_poll.c icmp_conn.c icmp_sendmsg.c